Added client cache setting
If client cache is set to true, the image will be used (from cache) when a browser request it. If the setting is set to false, the check if the cache is old will be done server side. That way you don't need to add timestamp to the url to refresh the user cache